requirements.txt
beard-portscan==0.1.5
Pillow==9.4.0
requests==2.28.2